requirements_ex_tables.txt
backcall==0.2.0
Brotli==1.0.9
chardet==4.0.0
click==8.0.3
colorama==0.4.4
dash==2.0.0
dash-bootstrap-components==1.0.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.0
Flask==2.0.2
Flask-Compress==1.10.1
ipdb==0.13.9
ipython==7.29.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.2
lxml==4.6.3
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
numexpr==2.7.3
numpy==1.21.3
pandas==1.3.4
parso==0.8.2
pickleshare==0.7.5
plotly==5.3.1
prompt-toolkit==3.0.21
pyarrow==6.0.0
Pygments==2.10.0
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
tenacity==8.0.1
toml==0.10.2
traitlets==5.1.1
wcwidth==0.2.5
Werkzeug==2.0.2
wfastcgi==3.0.0
